+++ app/imports/ui/layouts/App.jsx
@@ -8,14 +8,12 @@ import NavBar from '../components/NavBar';
 import Footer from '../components/Footer';
 import Landing from '../pages/Landing';
 import ListEvents from '../pages/ListEvents';
-import ListStuffAdmin from '../pages/ListStuffAdmin';
 import EditEvent from '../pages/EditEvent';
 import NotFound from '../pages/NotFound';
 import Signin from '../pages/Signin';
 import Signup from '../pages/Signup';
 import Signout from '../pages/Signout';
 import CreateEvent from '../pages/CreateEvent';
-import ICSGenerator from '../components/ICSGenerator';
 
 /** Top-level layout component for this application. Called in imports/startup/client/startup.jsx. */
 class App extends React.Component {
@@ -31,8 +29,6 @@ class App extends React.Component {
               <ProtectedRoute path="/list" component={ListEvents}/>
               <ProtectedRoute path="/add" component={CreateEvent}/>
               <ProtectedRoute path="/edit/:_id" component={EditEvent}/>
-              <ProtectedRoute path="/generate/:_id" component={ICSGenerator}/>
-                <AdminProtectedRoute path="/admin" component={ListStuffAdmin}/>
               <ProtectedRoute path="/signout" component={Signout}/>
               <Route component={NotFound}/>
             </Switch>
