+++ app/server/main.js
@@ -1,4 +1,3 @@
 import '/imports/startup/server/Accounts';
 import '/imports/startup/server/Publications';
 import '/imports/startup/server/Mongo';
-import '/imports/startup/server/Startup';
